move ICalendarWriter to ical package
